Move some more code to core.
