+++ worldguard-core/src/main/java/com/sk89q/worldguard/commands/task/RegionManagerReloader.java
@@ -17,7 +17,7 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.sk89q.worldguard.bukkit.commands.task;
+package com.sk89q.worldguard.commands.task;
 
 import com.sk89q.worldguard.protection.managers.RegionManager;
 import com.sk89q.worldguard.protection.managers.storage.StorageException;
